Removed and added parameters
removed average and compensated altitude calculations (not usable at all), removed altitude calculation according to pressure. Added day top elevation angle parameter.

+++ AVR64DD32-SUNFLOWER-WEATHER-STATION/Altitude.h
@@ -60,9 +60,9 @@
  * - `AVRG`: Average altitude (meters).
  */
 typedef struct { 
-    double UNCOMP; /**< Uncompensated altitude. */
-    double COMP;   /**< Compensated altitude. */
-    float AVRG;    /**< Average altitude. */
+    int16_t UNCOMP; /**< Uncompensated altitude. */
+    int16_t COMP;   /**< Compensated altitude. */
+    int16_t AVRG;    /**< Average altitude. */
 } Alt;
 
 /**
